Rename compiler to solc
