+++ crates/solidity/src/solc/solc_compiler.rs
