+++ crates/solidity/src/tests/ir_artifacts.rs
@@ -5,7 +5,7 @@
 
 use std::collections::BTreeMap;
 
-use crate::compiler::pipeline::Pipeline as SolcPipeline;
+use crate::solc::pipeline::Pipeline as SolcPipeline;
 
 #[test]
 fn yul() {
