+++ crates/solidity/src/tests/runtime_code.rs
@@ -4,7 +4,7 @@
 
 use std::collections::BTreeMap;
 
-use crate::compiler::pipeline::Pipeline as SolcPipeline;
+use crate::solc::pipeline::Pipeline as SolcPipeline;
 
 #[test]
 #[should_panic(expected = "runtimeCode is not supported")]
