+++ crates/solidity/src/tests/unsupported_opcodes.rs
@@ -4,7 +4,7 @@
 
 use std::collections::BTreeMap;
 
-use crate::compiler::pipeline::Pipeline as SolcPipeline;
+use crate::solc::pipeline::Pipeline as SolcPipeline;
 
 #[test]
 #[should_panic(expected = "The `CODECOPY` instruction is not supported")]
